refactor: all endpoints refactored
