+++ backend/api/media/__init__.py
@@ -1,0 +1,8 @@
+__all__ = [
+    "ServeProxy",
+    "GetProxyInfo",
+    "UploadMedia",
+]
+
+from .proxy import GetProxyInfo, ServeProxy
+from .upload import UploadMedia
