+++ backend/api/objects/__init__.py
@@ -1,0 +1,11 @@
+__all__ = [
+    "DeleteObjects",
+    "GetObjects",
+    "Operations",
+    "SetMetadata",
+]
+
+from .delete_objects import DeleteObjects
+from .get_objects import GetObjects
+from .operations import Operations
+from .set_metadata import SetMetadata
