--- backend/api/order_items/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["OrderItems"]
-
-from .order_items import OrderItems
